In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ functions_nested_loops/102-fibonacci.c
@@ -12,7 +12,7 @@
 int main(void)
 {
 	int i;
-	long long int fib[50];
+	long fib[50];
 
 	fib[0] = 1;
 	fib[1] = 2;
@@ -24,10 +24,10 @@ int main(void)
 
 	for (i = 0; i < 49; i++)
 	{
-		printf("%lld, ", fib[i]);
+		printf("%ld, ", fib[i]);
 	}
 
-	printf("%lld\n", fib[49]);
+	printf("%ld\n", fib[49]);
 
 	return (0);
 }
